notifications/serializers.py
"""
Serializers for notifications app.
"""
from rest_framework import serializers
from .models import Notification
from accounts.serializers import UserSerializer


class NotificationSerializer(serializers.ModelSerializer):
    """Serializer for Notification model."""
    sender = UserSerializer(read_only=True)

    class Meta:
        model = Notification
        fields = [
            'id', 'recipient', 'sender', 'notification_type',
            'message', 'is_read', 'created_at'
        ]
        read_only_fields = ['id', 'created_at']
