requirements.txt
# Django Core
Django==5.0.6
psycopg2-binary==2.9.9

# Django REST Framework
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
drf-spectacular==0.27.2

# GraphQL
graphene-django==3.2.2
django-graphql-jwt==0.4.0
graphene-file-upload==1.3.0

# Django Channels (WebSocket)
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Celery & Redis
celery==5.4.0
redis==5.0.5
django-redis==5.4.0

# Authentication
django-allauth==0.63.3
dj-rest-auth==6.0.0
social-auth-app-django==5.4.1

# CORS & Security
django-cors-headers==4.3.1
whitenoise==6.6.0

# Media & Storage
Pillow==10.3.0
django-storages==1.14.3
boto3==1.34.124

# Environment Variables
python-decouple==3.8

# Utilities
python-dateutil==2.9.0
pytz==2024.1

# Development & Testing
pytest==8.2.1
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
faker==25.2.0
django-extensions==3.2.3
